Fix a problem when cloning to an existing folder.
We should not ignore empty directories when indexing
the worktree before clone. This is because, if the repo contains
an empty folder, we'll fail to detect fast-forward relationship
between worktree and the repo.

After the merge, we should write out the index file, so that
existing empty directories won't be ignore.