[client] Support periodic sync.
This feature is controlled by setting "sync-interval" property of
a repo. If the property is not set or <= 0, it's disabled; if it's > 0,
sync the repo every specified seconds.

If periodic sync is set for a repo, auto change detection is
disabled. And in every sync, the worktree is scanned, and latest
changes from the server is downloaded.