setup: Do not try to generate seahub db when it exists
When using setup-seafile-mysql.sh it asks you if you want to use an
existing database or create a new one. Even when you choose the 'Use
existing' option, it tries to create the seahub DB, and then the
installer fails because it tries to create already existing tables.

Fix is to remember the answer to the previous choice and only create the
seahub DB if you are creating a new DB.