Wrap public declarations with G_BEGIN_DECLS/G_END_DECLS
Glib release 2.67.3 started conditionally using some C++ code[1] in the
public headers and it's no longer legal to include "glib.h" or anything
that includes it transitively from `extern "C"` block.

The solution recommended by the glib developers is to enclose all
declarations in a header within G_BEGIN_DECLS and G_END_DECLS as shown
in gobject developer manual[2]. Since G_BEGIN/END_DECLS expands to an
`extern "C"` block it's also acceptable to use it directly when "glib.h"
is not included in the current header file.

[1] https://gitlab.gnome.org/GNOME/glib/-/merge_requests/1715
[2] https://developer.gnome.org/gobject/stable/howto-gobject.html#howto-gobject-header